Always clear local state directories for non-incremental build

+++ libraries/tools/kotlin-gradle-plugin/src/main/kotlin/org/jetbrains/kotlin/gradle/tasks/tasksUtils.kt
@@ -3,6 +3,8 @@ package org.jetbrains.kotlin.gradle.tasks
 import org.gradle.api.GradleException
 import org.gradle.api.Task
 import org.jetbrains.kotlin.cli.common.ExitCode
+import org.jetbrains.kotlin.compilerRunner.GradleKotlinLogger
+import org.jetbrains.kotlin.compilerRunner.KotlinLogger
 import org.jetbrains.kotlin.gradle.plugin.kotlinDebug
 import org.jetbrains.kotlin.gradle.utils.outputsCompatible
 import java.io.File
@@ -18,23 +20,27 @@ fun throwGradleExceptionIfError(exitCode: ExitCode) {
     }
 }
 
-internal val <T : Task> T.outputDirectories: List<File>
-    get() = outputsCompatible.files.files.filter { it.isDirectory }
+internal fun <T : Task> T.localStateDirectories(): List<File> =
+    outputsCompatible.files.files.filter { it.isDirectory }
 
-internal fun <T : Task> T.clearOutputDirectories(reason: String? = null) {
-    logger.kotlinDebug {
+internal fun <T : Task> T.clearLocalStateDirectories(reason: String? = null) {
+    clearLocalStateDirectories(GradleKotlinLogger(logger), localStateDirectories(), reason)
+}
+
+internal fun clearLocalStateDirectories(log: KotlinLogger, localStateDirectories: List<File>, reason: String?) {
+    log.kotlinDebug {
         val suffix = reason?.let { " ($it)" }.orEmpty()
-        "Clearing output directories for task '$path'$suffix:"
+        "Clearing output directories$suffix:"
     }
-    val outputDirectories = outputDirectories
-    for (dir in outputDirectories) {
+    for (dir in localStateDirectories) {
+        if (!dir.exists()) continue
         when {
             dir.isDirectory -> {
                 dir.deleteRecursively()
                 dir.mkdirs()
-                logger.kotlinDebug { "  deleted $dir" }
+                log.kotlinDebug { "  deleted $dir" }
             }
-            else -> logger.kotlinDebug { "  skipping $dir (not a directory)" }
+            else -> log.kotlinDebug { "  skipping $dir (not a directory)" }
         }
     }
 }
